🎉 FEAT: 장바구니 조회 기능 장바구니 조회 기능 추가 장바구니에 담긴 물품이 없을 시, 에러가 아닌 빈 장바구니 반환 Product 숫자형 변수들 타입 변경

+++ src/main/java/com/supernova/fashionnova/cart/Cart.java
@@ -55,7 +55,7 @@ public class Cart {
     }
 
     // totalPrice 증가
-    public void incrementTotalPrice(Long price) {
+    public void incrementTotalPrice(int price) {
         this.totalPrice += price;
     }
 }
